+++ l10n_ve_accountant/models/__init__.py
@@ -13,5 +13,4 @@ from . import account_payment_term
 from . import tax_unit
 from . import res_currency
 from . import res_config_settings
-
-# from . import account_journal # ESTA HERENCIA NO SE IMPORTARÁ PORQUE ESTÁ GENERANDO ERROR, AL SOLUCIONAR, VOLVER A AGREGAR EN EN IMPORT
+from . import account_journal
