+++ l10n_ve_accountant/views/account_journal_views.xml
