--- l10n_ve_tax/models/__init__.py
@@ -1,5 +1,0 @@
-from . import account_tax
-from . import res_company
-from . import res_config_settings
-from . import account_move_line
-from . import account_journal
