--- l10n_ve_tax/models/res_company.py
@@ -1,58 +1,0 @@
-from odoo import api, fields, models
-
-
-class ResCompany(models.Model):
-    _inherit = "res.company"
-
-    unique_tax = fields.Boolean()
-    show_discount_on_moves = fields.Boolean()
-
-    exent_aliquot_sale = fields.Many2one("account.tax", domain=[("type_tax_use", "=", "sale")])
-    general_aliquot_sale = fields.Many2one("account.tax", domain=[("type_tax_use", "=", "sale")])
-    reduced_aliquot_sale = fields.Many2one("account.tax", domain=[("type_tax_use", "=", "sale")])
-    extend_aliquot_sale = fields.Many2one("account.tax", domain=[("type_tax_use", "=", "sale")])
-    not_show_reduced_aliquot_sale = fields.Boolean()
-    not_show_extend_aliquot_sale = fields.Boolean()
-
-    exent_aliquot_purchase = fields.Many2one(
-        "account.tax", domain=[("type_tax_use", "=", "purchase")]
-    )
-    general_aliquot_purchase = fields.Many2one(
-        "account.tax", domain=[("type_tax_use", "=", "purchase")]
-    )
-    reduced_aliquot_purchase = fields.Many2one(
-        "account.tax", domain=[("type_tax_use", "=", "purchase")]
-    )
-    extend_aliquot_purchase = fields.Many2one(
-        "account.tax", domain=[("type_tax_use", "=", "purchase")]
-    )
-    not_show_reduced_aliquot_purchase = fields.Boolean()
-
-    not_show_extend_aliquot_purchase = fields.Boolean()
-
-    config_deductible_tax = fields.Boolean()
-
-    no_deductible_general_aliquot_purchase = fields.Many2one(
-        "account.tax", domain=[("type_tax_use", "=", "purchase")]
-    )
-    no_deductible_reduced_aliquot_purchase = fields.Many2one(
-        "account.tax", domain=[("type_tax_use", "=", "purchase")]
-    )
-    no_deductible_extend_aliquot_purchase = fields.Many2one(
-        "account.tax", domain=[("type_tax_use", "=", "purchase")]
-    )
-
-    exent_aliquot_purchase_international = fields.Many2one("account.tax", domain=[("type_tax_use", "=", "purchase")])
-    general_aliquot_purchase_international = fields.Many2one("account.tax", domain=[("type_tax_use", "=", "purchase")])
-    reduced_aliquot_purchase_international = fields.Many2one("account.tax", domain=[("type_tax_use", "=", "purchase")])
-    extend_aliquot_purchase_international = fields.Many2one("account.tax", domain=[("type_tax_use", "=", "purchase")])
-
-    not_show_general_aliquot_purchase_international = fields.Boolean()
-
-    not_show_reduced_aliquot_purchase_international = fields.Boolean()
-
-    not_show_extend_aliquot_purchase_international = fields.Boolean()
-
-    not_show_international_purchase_in_book = fields.Boolean()
-
-
